Time: 0 ms (100%), Space: 28.7 MB (49.3%) - LeetHub

+++ 3232-find-if-digit-game-can-be-won/3232-find-if-digit-game-can-be-won.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool canAliceWin(vector<int>& nums) {
-        int singlesum= 0, doublesum = 0;
+        long long singlesum= 0, doublesum = 0;
         int n= nums.size(); 
         for(int x: nums){
             if(x<10) singlesum += x;
